+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Improved pilet injector to always return an array as metadata response
 - Removed dependency to `webpack-inject-plugin` in `piral-cli-webpack5`
 - Fixed issue with `getCurrentLayout` not returning default layout in SSR
+- Added options `--headers` and `--mode` to `pilet publish`
 
 ## 0.14.18 (March 15, 2022)